% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{checkArgs() helper function to check whether necessary arguments have been inputted in NEST() function}
\usage{
checkArgs(args, required.args, optional.args)
}
\arguments{
\item{args}{list of all parameters required in corresponding statFun() (function for getting location-specific association statistics T(v))}

\item{required.args}{- character vector with names of required arguments (i.e., what parameters *need* to be named in args list)}

\item{optional.args}{- character vector with anmes of optional arguments (i.e., what parameters will be set to their default versions if unspecified by user)}
}
\description{
checkArgs() helper function to check whether necessary arguments have been inputted in NEST() function
}
