% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichScore.R
\name{enrichScore}
\alias{enrichScore}
\title{function to calculate enrichment score given a map of brain-phenotype associations & network labels}
\usage{
enrichScore(stat.map, net.map, one.sided = TRUE, save.detail = FALSE)
}
\arguments{
\item{stat.map}{- vector (length V) of location (v)-specific associations}

\item{net.map}{- vector (also length V) consisting of all 0's, 1's. It is assumed that 1's denote locations inside the network of interest, and 0's denote all other locations}

\item{one.sided}{- default is TRUE, so that the running sum statistic increases in increments proportional to the magnitude of T(v) at a given location in the network; setting one.sided = FALSE will change the increment for locations in the network to increase by 1/(# locations in network)}

\item{save.detail}{- default is FALSE, so that the calculated enrichment score is the only information returned. it can be set to FALSE to return additional information used to calculate the enrichment score}
}
\description{
function to calculate enrichment score given a map of brain-phenotype associations & network labels
}
