\name{nest-package}
\alias{nest-package}
\alias{nest}
\docType{package}
\title{Analysis of non-equally spaced time series}
\description{Time series analysis for irregularly sampled time series: Correlation and Spectral analysis.}
\details{
\tabular{ll}{
Package: \tab nest\cr
Type: \tab Package\cr
Version: \tab 1.05.12\cr
Date: \tab 2016-02-08\cr
License: \tab GPL-2\cr
}

}
\author{Kira Rehfeld, Maintainer: Kira Rehfeld <krehfeld@awi.de>}

\references{
Rehfeld, K. and Kurths, J.: Similarity estimators for irregular and age-uncertain time series, Clim. Past, 10, 107-122, doi:10.5194/cp-10-107-2014, 2014.
Rehfeld, K., Marwan, N., Heitzig, J., and Kurths, J.: Comparison of correlation analysis techniques for irregularly sampled time series, Nonlin. Processes Geophys., 18, 389-404, doi:10.5194/npg-18-389-2011, 2011.
}

\keyword{package}


