% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalFun.R
\name{pvalFun}
\alias{pvalFun}
\title{pvalFun() helper function to calculate a p-value based on user-specified observed statistic and null distribution}
\usage{
pvalFun(obs, null.dist)
}
\arguments{
\item{obs}{- observed statistic (i.e. enrichment score)}

\item{null.dist}{- null distribution (i.e., enrichment scores under H0, obtained via permutation)}
}
\description{
pvalFun() helper function to calculate a p-value based on user-specified observed statistic and null distribution
}
