% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFun.gam.mvwald.R
\name{statFun.gam.mvwald}
\alias{statFun.gam.mvwald}
\title{statFun.gam.mvwald() will be called by NEST() if NEST argument statFun=="gam.mvwald"}
\usage{
statFun.gam.mvwald(
  X,
  dat,
  gam.formula,
  lm.formula,
  y.in.gam,
  y.in.lm,
  y.permute = NULL,
  n.cores = 1,
  seed = NULL,
  n.perm = 999,
  getNull = TRUE
)
}
\arguments{
\item{X}{n x p matrix of image measurements (n = number of subjects, p = number of image locations)}

\item{dat}{matrix including phenotype and any covariates. make sure all columns in dat are names}

\item{gam.formula}{user-specified formula for gam fit at each vertex (in each gam, outcome will be a different column of X and the right side of the formula should include variables included in dat matrix)}

\item{lm.formula}{user-specified formula for lm fit at each vertex (purpose is just to get a sign of the multivariate wald statistic obtained from the gam at each vertex, since the wald stat is always positive)}

\item{y.in.gam}{character specifying how phenotype variable can be identified in gam. if phenotype is included as a linear term in the gam formula, this could just be the variable name. if there's a smooth term for the phenotype, it may be something like s(y)}

\item{y.in.lm}{character specifying how phenotype variable can be identified in lm. again, might just be the name of the variable, but could be different (e.g., if testing an interaction term)}

\item{y.permute}{specify names of columns in dat matrix that should be permuted when getNull = TRUE}

\item{n.cores}{for parallelization, number of cores to use (default is 1 / no parallelization)}

\item{seed}{optional to set seed}

\item{n.perm}{number of permutations to conduct for inference. default is 999 (i.e. minimum p-value 1/1000)}

\item{getNull}{whether to obtain null distribution vs. just get observed map of statistics. default will be TRUE inside NEST function, but the statFun function will then be recursively called to get null distribution and getNull will then switch to FALSE}
}
\description{
statFun.gam.mvwald() will be called by NEST() if NEST argument statFun=="gam.mvwald"
}
