% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFun.lm.R
\name{statFun.lm}
\alias{statFun.lm}
\title{statFun.lm() will be called by NEST() if NEST argument statFun=="lm"}
\usage{
statFun.lm(
  X,
  y,
  Z = 1,
  type = "coef",
  n.cores = 1,
  seed = NULL,
  FL = FALSE,
  n.perm = 999,
  getNull = TRUE
)
}
\arguments{
\item{X}{n x p matrix (n = number of subjects, p = number of image locations)}

\item{y}{vector length n with phenotype measurements for each subject}

\item{Z}{vector (length n) or matrix (number of rows = n) with covariates to be adjusted for in lm. default Z = 1 assumes no covariates (i.e. Z=1 becomes placeholder for the intercept)}

\item{type}{type of statistic from lm() output to use. default is coefficient (will be the coefficient associated with phenotype variable y)}

\item{n.cores}{for parallelization, number of cores to use (default is 1 / no parallelization)}

\item{seed}{optional to set seed}

\item{FL}{whether to use Freedman-Lane approach to permutation. default is FALSE. this option may be preferable if phenotype/covariates are not independent}

\item{n.perm}{number of permutations to conduct for inference. default is 999 (i.e. minimum p-value 1/1000)}

\item{getNull}{whether to obtain null distribution vs. just get observed map of statistics. default will be TRUE inside NEST function, but the statFun function will then be recursively called to get null distribution and getNull will then switch to FALSE}
}
\description{
statFun.lm() will be called by NEST() if NEST argument statFun=="lm"
}
