\name{Nest-package}
\alias{Nest-package}
\alias{Nest}
\docType{package}
\title{
Study Random Genetic Drift
}
\description{
The \code{Nest} package allows you to simulate allele frequency trajectories under random genetic drift using the Wright-Fisher model and estimate the effective population size (Ne) applying different moment-based methods. Pool-seq data can be simulated by re-sampling allele count data. Additionally, allele frequency data can be read from file and Ne can be readily estimated for windows along the genome, or for whole chromosomes instead.
}
\details{

}
\author{
\packageAuthor{Nest}

Maintainer: \packageMaintainer{Nest}
}
\references{
Jonas, A., Taus, T., Kosiol, C., Schlötterer, C. & Futschik, A.: Estimating effective population size from temporal allele frequency changes in experimental evolution, manuscript in preparation
}
\keyword{ package }
\seealso{
\code{\link{checkSNP}}, \code{\link{estimateNe}}, \code{\link{estimateWndNe}}, \code{\link{read.sync}}, \code{\link{sample.alleles}} and \code{\link{wf.traj}}.
}
\examples{

}
