% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEST.R
\name{NEST}
\alias{NEST}
\title{NEST function}
\usage{
NEST(
  statFun,
  args,
  net.maps,
  one.sided = TRUE,
  stat.fun,
  n.cores = 1,
  seed = NULL,
  what.to.return = c("pval")
)
}
\arguments{
\item{statFun}{(to do: add descriptions)}

\item{args}{.....}

\item{net.maps}{.....}

\item{one.sided}{.....}

\item{stat.fun}{.....}

\item{n.cores}{.....}

\item{seed}{.....}

\item{what.to.return}{.....}
}
\description{
NEST function
}
